Added the text of the LGPLv2.1 to the repo.
This repo is intended to be used in conjunction with libgphoto2 and
gphoto2 but since it stands alone in its own repo as well, it should
include its own license.  Since libgphoto2 is licensed under the LGPL
and gphoto2 under the GPL, this repo gets the more liberal LGPL so it
can be distributed under both.